VARGAS-MAURICIO_SILVA-MARIA-DELOSANGELES/src/com/backend/service/OdontologoService.java
package com.backend.service;

import com.backend.DAO.IDao;
import com.backend.DAO.impl.OdontologoDaoH2;
import com.backend.entity.Odontologo;

import java.util.List;

public class OdontologoService {

    private IDao<Odontologo> odontologoIDao;

    public OdontologoService(IDao<Odontologo> odontologoIDao) {
        this.odontologoIDao = odontologoIDao;
    }

    public Odontologo guardarOdontologo(Odontologo odontologo){
        return odontologoIDao.guardar(odontologo);
    }

    public List<Odontologo> listarOdontologos(){
        return odontologoIDao.listarTodos();
    }

}
